requirements.txt
streamlit
python-dotenv
openai
PyPDF2
python-docx
beautifulsoup4